Add core processing components for CSV data handling in ForkliftCore
