+++ src/forklift/engine/config/__init__.py
@@ -1,0 +1,12 @@
+"""Configuration module for Forklift engine."""
+
+from .enums import HeaderMode, ExcessColumnMode
+from .import_config import ImportConfig
+from .processing_results import ProcessingResults
+
+__all__ = [
+    'HeaderMode',
+    'ExcessColumnMode',
+    'ImportConfig',
+    'ProcessingResults'
+]
