+++ src/forklift/engine/processors/base.py
@@ -1,0 +1,7 @@
+# processors/base.py
+from abc import ABC, abstractmethod
+
+class BaseProcessor(ABC):
+    @abstractmethod
+    def process(self, config: ImportConfig) -> ProcessingResults:
+        pass
